rename use case file
